requirements.txt
vkbottle~=4.1.5
environs~=9.3.5
aiosqlite~=0.17.0
aiofiles~=0.8.0
aiohttp~=3.8.1
disnake~=2.3.0
tortoise-orm==0.17.8
aerich==0.6.1